fix(review): preserve discovery for candidate seeds
